If all mcip files are already in place, exit run-wrf.sh early
The entire WRF/MCIP workflow is time consuming and resource intensive. If a daily run has already been completed and the MCIP results needed by the rest of the system are already available, we should be able to skip re-generating these files.

However, if any files are missing, or we explicitly want to regenerate them, the full script should still run.